feat(compiler): widen `ValueObj::Int` to i64
`Int` was an `i32`, so constant folding was asymmetric: positive results
were kept up to `u64::MAX` (as `Nat`) while anything below `i32::MIN` was
rejected. `1 - 10000000000` reported "cannot be evaluated" and the lexer
rejected `-3000000000` as an invalid literal.

Serialization follows the same rule `Nat` already did -- `TYPE_INT` when
the value fits, `TYPE_LONG` otherwise (verified on 3.10-3.14).

Two latent bugs fell out of the widening: `Int(i) == Nat(n)` compared
`i as u64`, so `Int(-1)` equalled `Nat(u64::MAX)`, and the mixed
`Int`/`Nat` comparisons truncated through `as i32`.

Phase 3 (integer range) of docs/const-eval-plan.md.

+++ crates/erg_compiler/context/eval.rs
@@ -2153,9 +2153,9 @@ impl Context {
                 _ => Err(uncomputable(&val)),
             },
             Neg => match &val {
-                ValueObj::Bool(b) => Ok(ValueObj::Int(-(*b as i32))),
+                ValueObj::Bool(b) => Ok(ValueObj::Int(-(*b as i64))),
                 // `Nat` is a `u64` but `Int` an `i32`, so negating can leave the range
-                ValueObj::Nat(n) => i32::try_from(*n)
+                ValueObj::Nat(n) => i64::try_from(*n)
                     .map(|n| ValueObj::Int(-n))
                     .map_err(|_| uncomputable(&val)),
                 ValueObj::Int(i) => i
